+++ PeterBoyajianU1M5Summative/src/test/java/com/trilogyed/PeterBoyajianU1M5Summative/doa/BookDaoTest.java
@@ -11,6 +11,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 //import java.time.LocalDate;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Date;
@@ -43,7 +44,7 @@ public class BookDaoTest {
         book.setIsbn("ISBN");
         book.setPublish_date(LocalDate.of(2019,1,1));
         book.setTitle("title");
-        book.setPrice(3.14);
+        book.setPrice(BigDecimal.valueOf(3.14));
         authors=authorDao.getAllAuthors();
         authors.forEach(a->authorDao.deleteAuthor(a.getAuthor_id()));
